Initial port of ufo_sources and opentype to new architecture
