+++ Lib/fontbakery/checks/cff.py
@@ -1,11 +1,7 @@
 from fontbakery.callable import check, condition
 from fontbakery.status import FAIL, PASS, WARN
 from fontbakery.message import Message
-
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
-
-profile_imports = ((".", ("shared_conditions",)),)
+from fontbakery.profiles.shared_conditions import ttFont, is_cff, is_cff2
 
 
 class CFFAnalysis:
