+++ Lib/fontbakery/checks/cmap.py
@@ -2,9 +2,6 @@ from fontbakery.callable import check
 from fontbakery.status import FAIL, PASS
 from fontbakery.message import Message
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
-
 
 @check(
     id="com.google.fonts/check/family/equal_unicode_encodings",
