+++ Lib/fontbakery/checks/dsig.py
