+++ Lib/fontbakery/checks/gdef.py
