+++ Lib/fontbakery/checks/glyf.py
@@ -4,8 +4,7 @@ from fontbakery.message import Message
 from fontbakery.callable import check
 from fontbakery.status import FAIL, PASS, WARN
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
+from fontbakery.profiles.shared_conditions import is_ttf
 
 
 @check(
