+++ Lib/fontbakery/checks/hhea.py
@@ -2,11 +2,6 @@ from fontbakery.callable import check
 from fontbakery.status import FAIL, PASS
 from fontbakery.message import Message
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
-
-profile_imports = [(".shared_conditions", ("glyph_metrics_stats", "is_ttf"))]
-
 
 @check(id="com.google.fonts/check/maxadvancewidth", proposal="legacy:check/073")
 def com_google_fonts_check_maxadvancewidth(ttFont):
