+++ Lib/fontbakery/checks/loca.py
@@ -2,10 +2,7 @@ from fontbakery.callable import check
 from fontbakery.status import PASS, FAIL
 from fontbakery.message import Message
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
-
-profile_imports = ((".", ("shared_conditions",)),)
+from fontbakery.profiles.shared_conditions import is_ttf
 
 
 @check(
