+++ Lib/fontbakery/checks/os2.py
@@ -2,14 +2,8 @@ from fontbakery.callable import check
 from fontbakery.status import FAIL, PASS, WARN, INFO, SKIP
 from fontbakery.message import Message
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
 
-profile_imports = [
-    (".shared_conditions", ("vmetrics",)),
-    (".googlefonts_conditions", ("RIBBI_ttFonts",)),
-]
-
+from fontbakery.profiles.googlefonts_conditions import RIBBI_ttFonts, style
 
 @check(
     id="com.google.fonts/check/family/panose_proportion", proposal="legacy:check/009"
