+++ Lib/fontbakery/checks/post.py
@@ -2,12 +2,6 @@ from fontbakery.callable import check
 from fontbakery.status import FAIL, PASS, WARN
 from fontbakery.message import Message
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
-
-profile_imports = [(".shared_conditions", ("is_ttf",))]
-
-
 @check(
     id="com.google.fonts/check/family/underline_thickness",
     rationale="""
