+++ Lib/fontbakery/checks/stat.py
@@ -5,11 +5,7 @@ from fontbakery.message import Message
 from fontbakery.status import FAIL, PASS, WARN, SKIP
 from fontbakery.utils import bullet_list
 
-# used to inform get_module_profile whether and how to create a profile
-from fontbakery.fonts_profile import profile_factory  # noqa:F401 pylint:disable=W0611
-
-profile_imports = ((".", ("shared_conditions",)),)
-
+from fontbakery.profiles.shared_conditions import is_variable_font, has_STAT_table, style
 
 @check(
     id="com.google.fonts/check/varfont/stat_axis_record_for_each_axis",
