+++ Lib/fontbakery/prelude.py
@@ -1,0 +1,4 @@
+from fontbakery.callable import check, condition, disable
+from fontbakery.message import Message
+from fontbakery.status import DEBUG, ERROR, FAIL, FATAL, INFO, PASS, SKIP, WARN
+from fontbakery.utils import exit_with_install_instructions
